WindowServer: Mark Vector of mutable references as mutable
The point of a reference type is to behave just like the referred-to
type. So, a Menu& should behave just like a Menu.

In this case, we had a const Vector. If it was a const Vector of Menu,
iterating over the Vector would only permit taking const references to
the individual Menues.

However, we have a const Vector of Menu&. The behavior should not
change. We should still only be permitted to take const references to
the individual Menues. Otherwise, we would be allowed to mutate the
individual Menues, which would mutate the elements of the const Vector.
This wouldn't modify the stored pointers, but it would modify the
objects that the references refer to. Since references should be
transparent, this should not be legal.

So it should be impossible to get mutable references into a const
Vector. Since we need mutable references in this case to call the
mutating member functions, we need to mark the Vector as mutable as
well.